feat: rename exercises file and tests file
