--- app/src/lib/musclesData.ts
@@ -1,3 +1,0 @@
-import type { MusclesDetails } from './types';
-import { Muscles } from './joints';
-
